move tests and pr comment
