+++ azext_edge/tests/edge/adr/test_asset_endpoint_profiles_int.py
@@ -8,8 +8,8 @@ import json
 import pytest
 from time import sleep
 from knack.log import get_logger
-from ....generators import generate_random_string
-from ....helpers import create_file, run
+from ...generators import generate_random_string
+from ...helpers import create_file, run
 
 logger = get_logger(__name__)
