+++ azext_edge/tests/edge/adr/test_namespaces_int.py
@@ -7,8 +7,8 @@
 from typing import List, Optional
 import pytest
 from knack.log import get_logger
-from ....generators import generate_random_string
-from ....helpers import run
+from ...generators import generate_random_string
+from ...helpers import run
 
 logger = get_logger(__name__)
