--- azext_edge/tests/edge/rpsaas/adr/__init__.py
@@ -1,5 +1,0 @@
-# coding=utf-8
-# ----------------------------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License. See License file in the project root for license information.
-# ----------------------------------------------------------------------------------------------
